refactor: factor out belts.rs
